feat: refactor topology analyzer with modular structure and enhanced functionality
- Introduced a new modular architecture for the topology analyzer by creating a `lib` directory containing shared utilities, filters, matching algorithms, and rendering functions.
- Moved common functions and constants from `ros2_topology_common.py` to `lib/common.py`, ensuring a single source of truth for shared resources.
- Implemented new filtering functions in `lib/filters.py` to manage node and topic visibility, improving report customization.
- Enhanced the matching algorithm in `lib/matching.py` to support more robust node identity recognition across snapshots.
- Added comprehensive report rendering capabilities in `lib/render.py`, allowing for detailed output of topology changes and node relationships.

These changes significantly improve the maintainability and extensibility of the topology analyzer, enabling better analysis and reporting of ROS 2 system configurations.
